requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.5.1
torch==2.0.1